+++ src/DictVectors/propagators.jl
@@ -103,106 +103,3 @@ function spawn_column!(column, o::OperatorMulPropagator, k, v)
         o.working_memory.style, column, o.operator, k, v, one(T), -one(T)
     )
 end
-
-###
-### Operator linear algebra operations
-###
-"""
-    mul!(y::PDVec, A::AbstractHamiltonian, x::PDVec[, w::PDWorkingMemory])
-
-Perform `y = A * x` in-place. The working memory `w` is required to facilitate
-threaded/distributed operations. If not passed a new instance will be allocated. `y` and `x`
-may be the same vector.
-"""
-function LinearAlgebra.mul!(y::PDVec, op::AbstractHamiltonian, x::PDVec, w=PDWorkingMemory(y))
-    prop = OperatorMulPropagator(op, w)
-    return mul!(y, prop, x)
-end
-
-function Base.:*(op::AbstractHamiltonian, t::PDVec)
-    prop = OperatorMulPropagator(op, t)
-    dst = similar(t, eltype(prop))
-    return mul!(dst, prop, t)
-end
-
-"""
-    dot(y::PDVec, A::AbstractHamiltonian, x::PDVec[, w::PDWorkingMemory])
-
-Perform `y ⋅ A ⋅ x`. The working memory `w` is required to facilitate threaded/distributed
-operations with non-diagonal `A`. If needed and not passed a new instance will be
-allocated. `A` can be replaced with a tuple of operators.
-"""
-function LinearAlgebra.dot(t::PDVec, op::AbstractHamiltonian, u::PDVec, w)
-    return dot(LOStructure(op), t, op, u, w)
-end
-function LinearAlgebra.dot(t::PDVec, op::AbstractHamiltonian, u::PDVec)
-    return dot(LOStructure(op), t, op, u)
-end
-function LinearAlgebra.dot(
-    ::IsDiagonal, t::PDVec, op::AbstractHamiltonian, u::PDVec, w=nothing
-)
-    T = promote_type(valtype(t), eltype(op), valtype(u))
-    return sum(pairs(u); init=zero(T)) do (k, v)
-        T(conj(t[k]) * diagonal_element(op, k) * v)
-    end
-end
-function LinearAlgebra.dot(
-    ::LOStructure, left::PDVec, op::AbstractHamiltonian, right::PDVec, w=nothing
-)
-    # First two cases: only one vector is distrubuted. Avoid shuffling things around
-    # by placing that one on the left to reduce the need for communication.
-    if !is_distributed(left) && is_distributed(right)
-        return dot(AdjointUnknown(), left, op, right)
-    elseif is_distributed(left) && !is_distributed(right)
-        return conj(dot(AdjointUnknown(), right, op', left))
-    end
-    # Other cases: both vectors distributed or not distributed. Put the shorter vector
-    # on the right as is done for regular DVecs.
-    if length(left) < length(right)
-        return conj(dot(AdjointUnknown(), right, op', left, w))
-    else
-        return dot(AdjointUnknown(), left, op, right, w)
-    end
-end
-# Default variant: also called from other LOStructures.
-function LinearAlgebra.dot(
-    ::AdjointUnknown, t::PDVec, op::AbstractHamiltonian, source::PDVec, w=nothing
-)
-    if is_distributed(t)
-        if isnothing(w)
-            target = copy_to_local!(PDWorkingMemory(t), t)
-        else
-            target = copy_to_local!(w, t)
-        end
-    else
-        target = t
-    end
-    return dot_from_right(target, op, source)
-end
-
-function dot_from_right(target, op, source::PDVec)
-    T = promote_type(valtype(target), valtype(source), eltype(op))
-    result = sum(pairs(source); init=zero(T)) do (k, v)
-        res = conj(target[k]) * diagonal_element(op, k) * v
-        for (k_off, v_off) in offdiagonals(op, k)
-            res += conj(target[k_off]) * v_off * v
-        end
-        res
-    end
-    return result::T
-end
-
-function LinearAlgebra.dot(t::PDVec, ops::Tuple, source::PDVec, w=nothing)
-    if is_distributed(t) && any(LOStructure(op) ≢ IsDiagonal() for op in ops)
-        if isnothing(w)
-            target = copy_to_local!(PDWorkingMemory(t), t)
-        else
-            target = copy_to_local!(w, t)
-        end
-    else
-        target = t
-    end
-    return map(ops) do op
-        dot_from_right(target, op, source)
-    end
-end
