+++ src/StochasticStyles/compression.jl
@@ -10,20 +10,20 @@ struct ThresholdCompression{T} <: CompressionStrategy
 end
 ThresholdCompression() = ThresholdCompression(1)
 
-function _threshold_compress!(t::ThresholdCompression, w, v)
-    for (add, val) in pairs(v)
+function _threshold_compress!(t::ThresholdCompression, w, pairs)
+    for (add, val) in pairs
         prob = abs(val) / t.threshold
         if prob < 1 # projection is only necessary if abs(val) < s.threshold
             val = ifelse(prob > rand(), t.threshold * sign(val), zero(val))
-            w[key] = val
         end
+        w[add] = val
     end
 end
 
 function compress!(t::ThresholdCompression, v)
     len_before = length(v)
     v_local = localpart(v)
-    _threshold_compress!(t, v_local, v_local)
+    _threshold_compress!(t, v_local, pairs(v_local))
     return (:len_before,), (len_before,)
 end
 
@@ -32,6 +32,6 @@ function compress!(t::ThresholdCompression, w, v)
     w_local = localpart(w)
     v_local = localpart(v)
     empty!(w_local)
-    _threshold_compress!(t, w_local, v_local)
+    _threshold_compress!(t, w_local, pairs(v_local))
     return (:len_before,), (len_before,)
 end
